# docs/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AstraWatch Unified API",
    "description": "Official OpenAPI 3.0 specification for AstraWatch autonomous telemetry, anomaly detection, incident orchestration, and K8s self-healing microservices.",
    "version": "1.0.0",
    "contact": {
      "name": "AstraWatch Engineering",
      "url": "https://astrawatch.io"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/v1",
      "description": "Collector Service (:8080)"
    },
    {
      "url": "http://localhost:8000/v1",
      "description": "Analyzer Service (:8000)"
    },
    {
      "url": "http://localhost:8082/api/v1",
      "description": "Orchestrator Service (:8082)"
    },
    {
      "url": "http://localhost:8081/healthz",
      "description": "K8s Operator Service (:8081 healthz)"
    },
    {
      "url": "http://localhost:8084",
      "description": "Realtime Gateway (:8084)"
    },
    {
      "url": "http://localhost:8085/api/v1",
      "description": "Payment Service (:8085)"
    }
  ],
...
  "paths": {
...
    "/api/v1/synthetics/checks": {
      "get": {
        "tags": [
          "Orchestrator: On-Call & Synthetics"
        ],
        "summary": "List Synthetic Checks",
        "description": "Lists all synthetic monitoring probes (HTTP, DNS, TCP).",
        "operationId": "osy1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "checks": [
                    {
                      "id": "chk-001",
                      "type": "HTTP",
                      "url": "https://api.example.com/health",
                      "interval": 60,
                      "lastResult": "OK"
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Orchestrator: On-Call & Synthetics"
        ],
        "summary": "Create Synthetic Check",
        "description": "Configures a new synthetic monitoring probe with assertions.",
        "operationId": "osy2",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "id": "chk-001",
                  "type": "HTTP",
                  "url": "https://api.example.com/health"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Check config",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "type": "HTTP",
                "url": "https://api.example.com/health",
                "interval": 60,
                "timeout": 10,
                "assertions": [
                  {
                    "type": "status_code",
                    "value": 200
                  }
                ]
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your Bearer JWT token issued by POST /api/v1/auth/login"
      }
    },
...
  }
}